fix: make stagedOnly work
Fix the stagedOnly option by switching from 'git add' to 'git update-index'. The -u flag for 'git add' does not in fact only update the index of staged files, but will also stage previously not-staged files. In fact, 'git add' does not have an option to do this. This is why the code now uses 'git update-index', which you can consider as a more powerful version of 'git add'. Evidenced, for a start, by the fact that it allows you to update the index only if the file was previously staged