chore: update husky install approach, remove pinst
Instead of installing husky with the `postinstall` hook, husky is now
installed with the `prepare` hook. The difference being that the former
also runs when this package is installed as a dependency. This was
previously prevented by `pinst` [1] (which removes the postinstall hook
before publishing to npm). Since `pinst` is no longer needed, it has
been removed as a dependency as well.

--
1. https://www.npmjs.com/package/pinst